refactor(host-application): modularize startup readiness and decouple runtime spawn config
