+++ apps/desktop/src-tauri/crates/host-application/src/app_service/opencode_runtime/startup_readiness/mod.rs
@@ -1,0 +1,11 @@
+mod policy;
+mod probe_runtime;
+mod wait_loop;
+
+pub(crate) use policy::{
+    OpencodeStartupReadinessPolicy, OpencodeStartupWaitReport, StartupCancelEpoch,
+};
+pub(crate) use wait_loop::wait_for_local_server_with_process;
+
+#[cfg(test)]
+pub(crate) use wait_loop::wait_for_local_server;
